I implemented CRUD operations for products

+++ client-processing/src/main/java/com/bankx/clientprocessing/repository/UserRepository.java
@@ -3,4 +3,4 @@ package com.bankx.clientprocessing.repository;
 import com.bankx.clientprocessing.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {}
+public interface UserRepository extends JpaRepository<UserEntity, Long> {}
